server/requirements.txt
Flask>=3.0
Flask-Login>=0.6
Flask-CORS>=4.0
SQLAlchemy>=2.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
marshmallow>=3.20
marshmallow-sqlalchemy>=1.0
bcrypt>=4.0
python-dotenv>=1.0
recipe-scrapers>=15.0
requests>=2.31
gunicorn>=23.0
psycopg2-binary>=2.9
